gke-cost-optimizer-agent/src/gke_cost_optimizer/cost_analyzer.py
"""
GCP Cost Analyzer - Integrates with GCP Billing API for cost analysis
"""

import asyncio
from datetime import datetime, timedelta
from typing import Dict, List, Any
from dataclasses import dataclass

from google.cloud import billing_v1
from google.cloud import monitoring_v3

@dataclass
class CostBreakdown:
    service: str
    sku: str
    amount: float
    currency: str
    usage_amount: float
    usage_unit: str

class CostAnalyzer:
    def __init__(self, project_id: str):
        self.project_id = project_id
        self.billing_client = billing_v1.CloudBillingClient()
        self.monitoring_client = monitoring_v3.MetricServiceClient()
        
        # GKE-related SKUs for cost analysis
        self.gke_skus = {
            'compute_engine': ['Compute Engine', 'Google Compute Engine'],
            'persistent_disk': ['Compute Engine Persistent Disk'],
            'load_balancer': ['Network Load Balancing', 'Cloud Load Balancing'],
            'gke_management': ['Google Kubernetes Engine'],
            'network': ['Compute Engine Network']
        }
    
    async def get_cluster_costs(self, days_back: int = 30) -> Dict[str, Any]:
        """Get comprehensive cost analysis for the cluster"""
        end_date = datetime.now()
        start_date = end_date - timedelta(days=days_back)
        
        cost_data = {
            'period': {
                'start': start_date.isoformat(),
                'end': end_date.isoformat(),
                'days': days_back
            },
            'total_cost': 0.0,
            'daily_costs': [],
            'service_breakdown': {},
            'resource_costs': {
                'compute': 0.0,
                'storage': 0.0,
                'network': 0.0,
                'management': 0.0
            },
            'cost_trends': {},
            'optimization_potential': {}
        }
        
        try:
            # Get billing account
            billing_account = await self._get_billing_account()
            if not billing_account:
                return self._get_mock_cost_data(cost_data)
            
            # Fetch cost data from billing API
            costs = await self._fetch_billing_data(billing_account, start_date, end_date)
            cost_data.update(costs)
            
            # Calculate optimization potential
            cost_data['optimization_potential'] = self._calculate_optimization_potential(cost_data)
            
        except Exception as e:
            print(f"Warning: Could not fetch billing data: {e}")
            return self._get_mock_cost_data(cost_data)
        
        return cost_data
    
    async def _get_billing_account(self) -> str:
        """Get the billing account for the project"""
        try:
            project_name = f"projects/{self.project_id}"
            project_billing = self.billing_client.get_project_billing_info(name=project_name)
            return project_billing.billing_account_name
        except Exception:
            return None
    
    async def _fetch_billing_data(self, billing_account: str, start_date: datetime, end_date: datetime) -> Dict[str, Any]:
        """Fetch actual billing data from GCP"""
        # Note: This requires billing export to BigQuery or Cloud Billing API access
        # For production, you'd query the billing export dataset
        
        costs = {
            'total_cost': 0.0,
            'service_breakdown': {},
            'daily_costs': []
        }
        
        # Mock implementation - in production, use BigQuery billing export
        # or Cloud Billing API with proper authentication
        
        return costs
    
    def _get_mock_cost_data(self, base_data: Dict[str, Any]) -> Dict[str, Any]:
        """Generate realistic mock cost data for demonstration"""
        import random
        
        # Generate daily costs for the period
        days = base_data['period']['days']
        daily_costs = []
        base_daily_cost = 150.0  # $150/day baseline
        
        for i in range(days):
            # Add some realistic variation
            daily_cost = base_daily_cost + random.uniform(-30, 50)
            daily_costs.append({
                'date': (datetime.now() - timedelta(days=days-i)).strftime('%Y-%m-%d'),
                'cost': round(daily_cost, 2)
            })
        
        total_cost = sum(day['cost'] for day in daily_costs)
        
        # Service breakdown
        service_breakdown = {
            'Compute Engine': {
                'cost': total_cost * 0.65,  # 65% compute
                'percentage': 65.0,
                'details': {
                    'n1-standard-4': total_cost * 0.40,
                    'n1-standard-2': total_cost * 0.25
                }
            },
            'Persistent Disk': {
                'cost': total_cost * 0.15,  # 15% storage
                'percentage': 15.0,
                'details': {
                    'pd-standard': total_cost * 0.10,
                    'pd-ssd': total_cost * 0.05
                }
            },
            'Load Balancing': {
                'cost': total_cost * 0.10,  # 10% networking
                'percentage': 10.0,
                'details': {
                    'forwarding_rules': total_cost * 0.06,
                    'data_processing': total_cost * 0.04
                }
            },
            'GKE Management': {
                'cost': total_cost * 0.05,  # 5% GKE management
                'percentage': 5.0
            },
            'Network': {
                'cost': total_cost * 0.05,  # 5% other network
                'percentage': 5.0
            }
        }
        
        # Resource costs
        resource_costs = {
            'compute': service_breakdown['Compute Engine']['cost'],
            'storage': service_breakdown['Persistent Disk']['cost'],
            'network': service_breakdown['Load Balancing']['cost'] + service_breakdown['Network']['cost'],
            'management': service_breakdown['GKE Management']['cost']
        }
        
        # Cost trends (comparing to previous period)
        cost_trends = {
            'total_change_percent': random.uniform(-10, 15),
            'compute_trend': random.uniform(-5, 20),
            'storage_trend': random.uniform(-15, 10),
            'network_trend': random.uniform(-8, 12)
        }
        
        return {
            **base_data,
            'total_cost': round(total_cost, 2),
            'daily_costs': daily_costs,
            'service_breakdown': service_breakdown,
            'resource_costs': resource_costs,
            'cost_trends': cost_trends
        }
    
    def _calculate_optimization_potential(self, cost_data: Dict[str, Any]) -> Dict[str, Any]:
        """Calculate potential cost savings opportunities"""
        total_cost = cost_data['total_cost']
        
        optimization_potential = {
            'total_potential_savings': 0.0,
            'opportunities': []
        }
        
        # Right-sizing opportunity (typically 20-30% savings)
        rightsizing_savings = total_cost * 0.25
        optimization_potential['opportunities'].append({
            'category': 'rightsizing',
            'description': 'Right-size over-provisioned resources',
            'potential_savings': rightsizing_savings,
            'confidence': 'high',
            'effort': 'medium'
        })
        
        # Spot instances opportunity (30-70% savings on compute)
        compute_cost = cost_data['resource_costs']['compute']
        spot_savings = compute_cost * 0.50
        optimization_potential['opportunities'].append({
            'category': 'spot_instances',
            'description': 'Use preemptible/spot instances for non-critical workloads',
            'potential_savings': spot_savings,
            'confidence': 'medium',
            'effort': 'high'
        })
        
        # Storage optimization (10-20% savings)
        storage_cost = cost_data['resource_costs']['storage']
        storage_savings = storage_cost * 0.15
        optimization_potential['opportunities'].append({
            'category': 'storage_optimization',
            'description': 'Optimize storage classes and remove unused volumes',
            'potential_savings': storage_savings,
            'confidence': 'high',
            'effort': 'low'
        })
        
        # Load balancer optimization (remove unused LBs)
        network_cost = cost_data['resource_costs']['network']
        lb_savings = network_cost * 0.30
        optimization_potential['opportunities'].append({
            'category': 'load_balancer_cleanup',
            'description': 'Remove unused load balancers and optimize networking',
            'potential_savings': lb_savings,
            'confidence': 'high',
            'effort': 'low'
        })
        
        optimization_potential['total_potential_savings'] = sum(
            opp['potential_savings'] for opp in optimization_potential['opportunities']
        )
        
        return optimization_potential
    
    def get_cost_per_resource(self, resource_type: str, resource_name: str) -> Dict[str, Any]:
        """Get cost breakdown for a specific resource"""
        # This would integrate with GCP's resource-level billing
        # For now, return estimated costs based on resource type
        
        cost_estimates = {
            'node': {
                'n1-standard-1': 24.27,  # Monthly cost
                'n1-standard-2': 48.55,
                'n1-standard-4': 97.09,
                'e2-standard-2': 40.32,
                'e2-standard-4': 80.64
            },
            'disk': {
                'pd-standard': 0.04,  # Per GB per month
                'pd-ssd': 0.17,
                'pd-balanced': 0.10
            },
            'load_balancer': {
                'forwarding_rule': 18.25,  # Monthly cost
                'data_processing': 0.008  # Per GB
            }
        }
        
        return cost_estimates.get(resource_type, {})
    
    def calculate_savings_projection(self, current_cost: float, optimization_percentage: float, months: int = 12) -> Dict[str, Any]:
        """Calculate projected savings over time"""
        monthly_savings = current_cost * (optimization_percentage / 100)
        
        return {
            'monthly_savings': round(monthly_savings, 2),
            'annual_savings': round(monthly_savings * 12, 2),
            'projected_savings': round(monthly_savings * months, 2),
            'roi_months': months,
            'optimization_percentage': optimization_percentage
        }